langchain-doc-qa-bot-langchain/requirements.txt
 langchain
openai
faiss-cpu
chromadb
python-dotenv
pypdf
tiktoken
streamlit

